+++ packages/backend/routes/coverage.ts
@@ -1,0 +1,10 @@
+import type { FastifyPluginAsync } from 'fastify';
+
+export const coverage: FastifyPluginAsync = async (
+  fastify,
+  opts
+): Promise<void> => {
+  fastify.get('/', (request, reply) => {
+    return global.__coverage__ ? { coverage: global.__coverage__ } : null;
+  });
+};